telebottest05/requirements.txt
python-telegram-bot==20.0
python-dotenv==0.21.0
requests==2.31.0
